+++ lib/Perl/Tidy.pm
@@ -3789,8 +3789,7 @@ sub generate_options {
     $add_option->( 'long-block-line-count',           'lbl',  '=i' );
     $add_option->( 'maximum-consecutive-blank-lines', 'mbl',  '=i' );
     $add_option->( 'keep-old-blank-lines',            'kbl',  '=i' );
-
-    $add_option->( 'keep-old-blank-lines-exclusion-list', 'kblxl', '=s' );
+    $add_option->( 'keep-old-blank-lines-exceptions', 'kblx', '=s' );
 
     $add_option->( 'keyword-group-blanks-list',         'kgbl', '=s' );
     $add_option->( 'keyword-group-blanks-size',         'kgbs', '=s' );