+++ t/snippets31.t
@@ -14,6 +14,12 @@
 #11 c446.def
 #12 pot.def
 #13 pot.pot1
+#14 kblx.def
+#15 kblx.kblx1
+#16 kblx.kblx2
+#17 kblx.kblx3
+#18 kblx.kblx4
+#19 kblx.kblx5
 
 # To locate test #13 you can search for its name or the string '#13'
 
@@ -40,6 +46,51 @@ BEGIN {
 ----------
         'def'   => "",
         'ils'   => "-nils -bos",
+        'kblx1' => <<'----------',
+-nbbc
+-blbs=0
+-blbp=0
+-nbbb
+-blao=0
+-blbc=0
+-kblx='b{' # inverse of -bbb
+----------
+        'kblx2' => <<'----------',
+-nbbc
+-blbs=0
+-blbp=0
+-nbbb
+-blao=0
+-blbc=0
+-kblx='b}' # inverse of -blbc
+----------
+        'kblx3' => <<'----------',
+-nbbc
+-blbs=0
+-blbp=0
+-nbbb
+-blao=0
+-blbc=0
+-kblx='{b' # inverse of -blao
+----------
+        'kblx4' => <<'----------',
+-nbbc
+-blbs=0
+-blbp=0
+-nbbb
+-blao=0
+-blbc=0
+-kblx='}b' # not an inverse
+----------
+        'kblx5' => <<'----------',
+-nbbc
+-blbs=0
+-blbp=0
+-nbbb
+-blao=0
+-blbc=0
+-kblx='bp' # inverse of -blbp
+----------
         'mutt1' => <<'----------',
 -mutt='q*'
 ----------
@@ -89,6 +140,45 @@ push @paths, qw(
         'ils' => <<'----------',
 $z = sqrt( $x**2 + $y**2 )
 ;
+----------
+
+        'kblx' => <<'----------',
+package A::B;
+
+sub write_line {
+
+    my ( $self, $line ) = @_;
+
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+
+}
+
+package C::D;
+
+sub dump_verbatim {
+
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+
+}    ## static side comment
+
+# coment before side comment
+
+# comment after blank and comment
 ----------
 
         'mutt' => <<'----------',
@@ -297,6 +387,255 @@ sub bla_p( $value = 42 ) {
       ->catch( sub { warn shift } );
 }
 #13...........
+        },
+
+        'kblx.def' => {
+            source => "kblx",
+            params => "def",
+            expect => <<'#14...........',
+package A::B;
+
+sub write_line {
+
+    my ( $self, $line ) = @_;
+
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+
+}
+
+package C::D;
+
+sub dump_verbatim {
+
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+
+}    ## static side comment
+
+# coment before side comment
+
+# comment after blank and comment
+#14...........
+        },
+
+        'kblx.kblx1' => {
+            source => "kblx",
+            params => "kblx1",
+            expect => <<'#15...........',
+package A::B;
+
+sub write_line {
+
+    my ( $self, $line ) = @_;
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+
+}
+
+package C::D;
+
+sub dump_verbatim {
+
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+
+}    ## static side comment
+
+# coment before side comment
+
+# comment after blank and comment
+#15...........
+        },
+
+        'kblx.kblx2' => {
+            source => "kblx",
+            params => "kblx2",
+            expect => <<'#16...........',
+package A::B;
+
+sub write_line {
+
+    my ( $self, $line ) = @_;
+
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+}
+
+package C::D;
+
+sub dump_verbatim {
+
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+}    ## static side comment
+
+# coment before side comment
+
+# comment after blank and comment
+#16...........
+        },
+
+        'kblx.kblx3' => {
+            source => "kblx",
+            params => "kblx3",
+            expect => <<'#17...........',
+package A::B;
+
+sub write_line {
+    my ( $self, $line ) = @_;
+
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+
+}
+
+package C::D;
+
+sub dump_verbatim {
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+
+}    ## static side comment
+
+# coment before side comment
+
+# comment after blank and comment
+#17...........
+        },
+
+        'kblx.kblx4' => {
+            source => "kblx",
+            params => "kblx4",
+            expect => <<'#18...........',
+package A::B;
+
+sub write_line {
+
+    my ( $self, $line ) = @_;
+
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+
+}
+package C::D;
+
+sub dump_verbatim {
+
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+
+}    ## static side comment
+# coment before side comment
+
+# comment after blank and comment
+#18...........
+        },
+
+        'kblx.kblx5' => {
+            source => "kblx",
+            params => "kblx5",
+            expect => <<'#19...........',
+package A::B;
+
+sub write_line {
+
+    my ( $self, $line ) = @_;
+
+    if ( defined($line) ) {
+        $self->write_line($line);
+    }
+    return;
+
+}
+package C::D;
+
+sub dump_verbatim {
+
+    my $self = shift;
+
+    # block comments
+    # block comments
+
+    my $rlines = $self->[_rlines_];
+
+    foreach my $line ( @{$rlines} ) {
+        my $input_line = $line->{_line_text};
+        $self->write_unindented_line($input_line);
+    }
+
+    return;
+
+}    ## static side comment
+
+# coment before side comment
+
+# comment after blank and comment
+#19...........
         },
     };